[Wisp] Clear pending exceptions of SOF/OOM when polling happens.
Summary: If a polling page happens in an OSR method which optimized the rethrow handler in C2-compiled method, and in the mean time a wisp preempt happens and throws an SOF/OOM exception, jvm will crash. We directly clear the pending ex to fix this "MISSING EXCEPTION HANDLER" problem.

Test Plan: TestPreempt.java